quiz-tutor-bot/requirements.txt
# FastAPI and server dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# LLM and AI dependencies
openai==1.54.3
anthropic==0.39.0

# Vector database
chromadb==0.5.20
sentence-transformers==3.3.1

# Data processing
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3

# CORS and middleware
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email parsing
email-validator==2.2.0
